requirements.txt
pigpio
matplotlib
google-cloud-speech
google-cloud-storage
openai
scikit-learn
psycopg2
